refactor: move vip to corresponding network folder
